feat: private functions and testing requires
